requirements.txt
fastapi
uvicorn
opencv-python
numpy
onnxruntime
pillow
pydload
python-dotenv
python-multipart